Exclude transits for items miscataloged at the system level.
While not a common issue, if any items that are miscataloged
with an owning library of a PINES system rather than branch,
the consolidation step fails because there's more than one
row returned from a subquery expecting a single value.  Explicitly
exclude any items owned by units with a parent_ou of 1 (PINES).